move client/unversioned/fake.go to its own package
