+++ pkg/client/unversioned/helper_test.go
@@ -17,16 +17,11 @@ limitations under the License.
 package unversioned
 
 import (
-	"bytes"
-	"encoding/json"
-	"io"
-	"io/ioutil"
 	"net/http"
 	"reflect"
 	"strings"
 	"testing"
 
-	"k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/api/testapi"
 )
 
@@ -377,86 +372,5 @@ func TestSetKubernetesDefaultsUserAgent(t *testing.T) {
 	}
 	if !strings.Contains(config.UserAgent, "kubernetes/") {
 		t.Errorf("no user agent set: %#v", config)
-	}
-}
-
-func objBody(object interface{}) io.ReadCloser {
-	output, err := json.MarshalIndent(object, "", "")
-	if err != nil {
-		panic(err)
-	}
-	return ioutil.NopCloser(bytes.NewReader([]byte(output)))
-}
-
-func TestNegotiateVersion(t *testing.T) {
-	tests := []struct {
-		name, version, expectedVersion string
-		serverVersions                 []string
-		clientVersions                 []string
-		config                         *Config
-		expectErr                      bool
-	}{
-		{
-			name:            "server supports client default",
-			version:         "version1",
-			config:          &Config{},
-			serverVersions:  []string{"version1", testapi.Default.Version()},
-			clientVersions:  []string{"version1", testapi.Default.Version()},
-			expectedVersion: "version1",
-			expectErr:       false,
-		},
-		{
-			name:            "server falls back to client supported",
-			version:         testapi.Default.Version(),
-			config:          &Config{},
-			serverVersions:  []string{"version1"},
-			clientVersions:  []string{"version1", testapi.Default.Version()},
-			expectedVersion: "version1",
-			expectErr:       false,
-		},
-		{
-			name:            "explicit version supported",
-			version:         "",
-			config:          &Config{Version: testapi.Default.Version()},
-			serverVersions:  []string{"version1", testapi.Default.Version()},
-			clientVersions:  []string{"version1", testapi.Default.Version()},
-			expectedVersion: testapi.Default.Version(),
-			expectErr:       false,
-		},
-		{
-			name:            "explicit version not supported",
-			version:         "",
-			config:          &Config{Version: testapi.Default.Version()},
-			serverVersions:  []string{"version1"},
-			clientVersions:  []string{"version1", testapi.Default.Version()},
-			expectedVersion: "",
-			expectErr:       true,
-		},
-	}
-	codec := testapi.Default.Codec()
-
-	for _, test := range tests {
-		fakeClient := &FakeRESTClient{
-			Codec: codec,
-			Resp: &http.Response{
-				StatusCode: 200,
-				Body:       objBody(&api.APIVersions{Versions: test.serverVersions}),
-			},
-			Client: HTTPClientFunc(func(req *http.Request) (*http.Response, error) {
-				return &http.Response{StatusCode: 200, Body: objBody(&api.APIVersions{Versions: test.serverVersions})}, nil
-			}),
-		}
-		c := NewOrDie(test.config)
-		c.Client = fakeClient.Client
-		response, err := NegotiateVersion(c, test.config, test.version, test.clientVersions)
-		if err == nil && test.expectErr {
-			t.Errorf("expected error, got nil for [%s].", test.name)
-		}
-		if err != nil && !test.expectErr {
-			t.Errorf("unexpected error for [%s]: %v.", test.name, err)
-		}
-		if response != test.expectedVersion {
-			t.Errorf("expected version %s, got %s.", test.expectedVersion, response)
-		}
 	}
 }
